client/requirements.txt
websockets==11.0.3
pyperclip==1.8.2
pillow==10.0.0
pywin32==306